refactor: harden BCP47 functions
