+++ komga/src/main/kotlin/org/gotson/komga/infrastructure/validation/BCP47.kt
@@ -1,9 +1,9 @@
 package org.gotson.komga.infrastructure.validation
 
-import com.ibm.icu.util.ULocale
 import jakarta.validation.Constraint
 import jakarta.validation.ConstraintValidator
 import jakarta.validation.ConstraintValidatorContext
+import org.gotson.komga.domain.model.BCP47TagValidator
 import kotlin.reflect.KClass
 
 @Constraint(validatedBy = [BCP47Validator::class])
@@ -23,19 +23,4 @@ class BCP47Validator : ConstraintValidator<BCP47, String> {
   }
 }
 
-object BCP47TagValidator {
-  private val languages by lazy { ULocale.getISOLanguages().toSet() }
 
-  fun isValid(value: String): Boolean {
-    return ULocale.forLanguageTag(value).let {
-      it.language.isNotBlank() && languages.contains(it.language)
-    }
-  }
-
-  fun normalize(value: String): String =
-    try {
-      ULocale.forLanguageTag(value).toLanguageTag()
-    } catch (e: Exception) {
-      ""
-    }
-}
